net,vm import: added fixtures to import a VM from external cluster (#2970)
Adding fixtures to import a VM from external cluster + simple source
cluster provider to operate VMs.

The following functionality has been added:
1) MTV operator sanity check: if mtv labled tests are collected, we need
to check whether the MTV operator is installed in the cluster.
2) source cluster provider to operate VMs: on the first iteration, it is
enough to have a minimal provider that will allow us to power on and off
the prepared VM for migration.
3) fixtures to import external source VM import to Openshift UDN
namespace: fixtures provide the ability to import a target VM from an
external cluster to the Openshift UDN namespace. Also added an initial
test that waits for the migration to complete. But the test has been
quarantined due to the migration speed issue, which will be handled
separately.

Signed-off-by: Sergei Volkov <sevolkov@redhat.com>